move CRC16 to own class
